refactor and optimize front-end table of test cases
