+++ test-server/src/main/java/com/clsaa/dop/server/test/interceptor/RequestUserInterceptor.java
@@ -1,21 +1,14 @@
 package com.clsaa.dop.server.test.interceptor;
 
-import com.clsaa.dop.server.test.util.UserUtils;
+import com.clsaa.dop.server.test.manager.UserManager;
 import com.clsaa.rest.result.bizassert.BizAssert;
-import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.actuate.metrics.web.servlet.LongTaskTimingHandlerInterceptor;
-import org.springframework.http.HttpRequest;
-import org.springframework.http.client.ClientHttpRequestExecution;
-import org.springframework.http.client.ClientHttpRequestInterceptor;
-import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
 
 import static org.apache.commons.lang3.StringUtils.isEmpty;
 
@@ -36,12 +29,12 @@ public class RequestUserInterceptor implements HandlerInterceptor {
         if (isEmpty(userId) && loginNeeded) {
             BizAssert.justUnauthorized("You should login first before accessing the system!");
         }
-        UserUtils.setUserId(userId);
+        UserManager.setUserId(userId);
         return true;
     }
 
     @Override
     public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
-        UserUtils.removeThreadUserId();
+        UserManager.removeThreadUserId();
     }
 }
