+++ test-server/src/main/java/com/clsaa/dop/server/test/mapper/param2po/InterfaceStagePoMapper.java
@@ -9,12 +9,11 @@ import com.clsaa.dop.server.test.model.po.WaitOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
-import static com.clsaa.dop.server.test.util.UserUtils.dateAndUser;
+import static com.clsaa.dop.server.test.manager.UserManager.dateAndUser;
 
 /**
  * @author xihao
