+++ test-server/src/main/java/com/clsaa/dop/server/test/mapper/param2po/TestStepPoMapper.java
@@ -5,10 +5,9 @@ import com.clsaa.dop.server.test.model.param.TestStepParam;
 import com.clsaa.dop.server.test.model.po.TestStep;
 import org.springframework.stereotype.Component;
 
-import java.time.LocalDateTime;
 import java.util.Optional;
 
-import static com.clsaa.dop.server.test.util.UserUtils.dateAndUser;
+import static com.clsaa.dop.server.test.manager.UserManager.dateAndUser;
 
 /**
  * @author xihao
