+++ test-server/src/main/java/com/clsaa/dop/server/test/mapper/po2dto/InterfaceCaseDtoMapper.java
@@ -4,6 +4,7 @@ import com.clsaa.dop.server.test.mapper.AbstractCommonServiceMapper;
 import com.clsaa.dop.server.test.model.dto.InterfaceCaseDto;
 import com.clsaa.dop.server.test.model.dto.InterfaceStageDto;
 import com.clsaa.dop.server.test.model.po.InterfaceCase;
+import com.clsaa.dop.server.test.manager.UserManager;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -34,7 +35,7 @@ public class InterfaceCaseDtoMapper extends AbstractCommonServiceMapper<Interfac
 
     @Override
     public Optional<InterfaceCaseDto> convert(InterfaceCase source) {
-        return super.convert(source).map(fillStages(source));
+        return super.convert(source).map(fillStages(source)).map(fillUser());
     }
 
     private Function<InterfaceCaseDto, InterfaceCaseDto> fillStages(InterfaceCase interfaceCase) {
@@ -42,6 +43,14 @@ public class InterfaceCaseDtoMapper extends AbstractCommonServiceMapper<Interfac
             List<InterfaceStageDto> interfaceStageDtos = interfaceStageDtoMapper.convert(interfaceCase.getStages());
             dto.setStages(interfaceStageDtos);
             return dto;
+        };
+    }
+
+    private Function<InterfaceCaseDto, InterfaceCaseDto> fillUser() {
+        return interfaceCaseDto -> {
+            Long cuserId = interfaceCaseDto.getCuser();
+            interfaceCaseDto.setCreateUserName(UserManager.getUserName(cuserId));
+            return interfaceCaseDto;
         };
     }
 }
