+++ test-server/src/main/java/com/clsaa/dop/server/test/mapper/po2dto/InterfaceExecuteLogDtoMapper.java
@@ -1,22 +1,17 @@
 package com.clsaa.dop.server.test.mapper.po2dto;
 
 import com.clsaa.dop.server.test.enums.Stage;
-import com.clsaa.dop.server.test.feign.UserInterface;
 import com.clsaa.dop.server.test.mapper.AbstractCommonServiceMapper;
 import com.clsaa.dop.server.test.model.dto.InterfaceExecuteLogDto;
 import com.clsaa.dop.server.test.model.dto.OperationExecuteLogDto;
-import com.clsaa.dop.server.test.model.dto.User;
 import com.clsaa.dop.server.test.model.po.InterfaceExecuteLog;
 import com.clsaa.dop.server.test.model.po.OperationExecuteLog;
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
+import com.clsaa.dop.server.test.manager.UserManager;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.*;
-import java.util.concurrent.ExecutionException;
 import java.util.stream.Collectors;
 
 /**
@@ -30,22 +25,6 @@ public class InterfaceExecuteLogDtoMapper extends AbstractCommonServiceMapper<In
 
     @Autowired
     private OperationLogDtoMapper operationLogDtoMapper;
-
-    @Autowired
-    private UserInterface userInterface;
-
-    private LoadingCache<Long,String> userNameCache = CacheBuilder.newBuilder().maximumSize(10000L).build(new CacheLoader<Long, String>() {
-        @Override
-        public String load(Long key) throws Exception {
-            User user = userInterface.getUserById(key);
-            if (user == null) {
-                log.error("[Get User Info From Feign]: error! invalid user id: {}", key);
-                return "Invalid UserId " + key;
-//                BizAssert.justInvalidParam("[Get User Info From Feign]: error! invalid user id: {}" + key);
-            }
-            return user.getName();
-        }
-    });
 
     @Override
     public Class<InterfaceExecuteLog> getSourceClass() {
@@ -75,11 +54,7 @@ public class InterfaceExecuteLogDtoMapper extends AbstractCommonServiceMapper<In
 
             // user name
             Long cuser = interfaceExecuteLogDto.getCuser();
-            try {
-                interfaceExecuteLogDto.setCreateUserName(userNameCache.get(cuser));
-            } catch (ExecutionException e) {
-                log.error("Get user name error! userId: {}", cuser, e);
-            }
+            interfaceExecuteLogDto.setCreateUserName(UserManager.getUserName(cuser));
 
             return interfaceExecuteLogDto;
         });
